rename qwen3 asr ci helpers
