+++ tests/test_model/test_qwen3_omni_mmsu_talker_ci.py
@@ -26,7 +26,7 @@ from pathlib import Path
 
 import pytest
 
-pytest_plugins = ["tests.test_model.omni_whisper_wer_utils"]
+pytest_plugins = ["tests.test_model.qwen3_asr_wer_utils"]
 
 from benchmarks.dataset.mmsu import load_mmsu_samples
 from benchmarks.dataset.prepare import DATASETS
@@ -193,7 +193,7 @@ def test_mmsu_talker_accuracy_and_speed(
 @pytest.mark.benchmark
 def test_mmsu_talker_wer(
     wer_eval_artifacts: _TalkerEvalArtifacts,
-    omni_whisper_wer_router: ManagedRouterHandle,
+    qwen3_asr_wer_router: ManagedRouterHandle,
 ) -> None:
     """Transcribe saved talker audio after the inference server is stopped."""
     wer = compute_text_audio_consistency_from_records(
@@ -202,7 +202,7 @@ def test_mmsu_talker_wer(
         ASR_DEVICE,
         audio_dir=wer_eval_artifacts.audio_dir,
         text_key="raw_response",
-        asr_router_port=omni_whisper_wer_router.port,
+        asr_router_port=qwen3_asr_wer_router.port,
     )
     print_wer_summary(wer["summary"], "qwen3-omni")
     persist_wer_in_benchmark_results(
