+++ tests/test_model/test_qwen3_omni_tts_ci.py
@@ -18,7 +18,7 @@ from pathlib import Path
 
 import pytest
 
-pytest_plugins = ["tests.test_model.omni_whisper_wer_utils"]
+pytest_plugins = ["tests.test_model.qwen3_asr_wer_utils"]
 
 from benchmarks.dataset.prepare import DATASETS, download_dataset
 from benchmarks.eval.benchmark_omni_seedtts import (
@@ -35,7 +35,7 @@ from tests.test_model.omni_router_utils import (
     print_worker_snapshot,
     router_get_json,
 )
-from tests.test_model.omni_whisper_wer_utils import wait_for_gpu_memory_release
+from tests.test_model.qwen3_asr_wer_utils import wait_for_gpu_memory_release
 from tests.utils import (
     MetricCheckCollector,
     apply_mos_slack,
@@ -458,12 +458,12 @@ def test_voice_cloning_non_streaming(
 def test_voice_cloning_wer(
     wer_audio_dir: str,
     dataset_repo: str,
-    omni_whisper_wer_router: ManagedRouterHandle,
+    qwen3_asr_wer_router: ManagedRouterHandle,
 ) -> None:
     results = _run_wer_transcribe(
         dataset_repo,
         wer_audio_dir,
-        asr_router_port=omni_whisper_wer_router.port,
+        asr_router_port=qwen3_asr_wer_router.port,
     )
     print_wer_summary(results["summary"], "qwen3-omni")
     checks = MetricCheckCollector("Qwen3-Omni voice-cloning WER")
@@ -474,7 +474,7 @@ def test_voice_cloning_wer(
         collector=checks,
     )
     checks.assert_all()
-    print_log_tail("asr_wer_router", omni_whisper_wer_router.log_file)
+    print_log_tail("asr_wer_router", qwen3_asr_wer_router.log_file)
 
 
 @pytest.mark.benchmark
